Moving ExportView and models out of the BagView namespace (this is to be shared with BagView and ResourceViews)
