+++ SanBag/Views/ExportView.xaml.cs
@@ -1,25 +1,8 @@
-﻿using LibSanBag;
-using SanBag.Commands;
-using SanBag.ViewModels;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.IO;
-using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
+﻿using System.ComponentModel;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using SanBag.ViewModels.BagViewModels;
+using SanBag.ViewModels;
 
-namespace SanBag.Views.BagViews
+namespace SanBag.Views
 {
     /// <summary>
     /// Interaction logic for ExportWindow.xaml
